Update index.html
In this Hamburger icon is in left side make it right in the mobile view. 
In the mobile view logo in the hero is more toward right, place it in a standard place towards left. 
Hamburger is not clickable in mobile view In mobile view when clicked on hamburger and selected products no dropdown is coming and not able to select any products. 
When clicked on hamburger in mobile view, only three nav header are shown and rest three are getting overlapped.